% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aep_nodes.R
\name{aep_node_human_cols}
\alias{aep_node_human_cols}
\title{Columns Owned by the Human}
\usage{
aep_node_human_cols()
}
\value{
A character vector of column names.
}
\description{
Never overwritten by \code{\link[=scaffold_aep_nodes]{scaffold_aep_nodes()}}.
}
