% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_themes.R
\docType{data}
\name{arctic_circle_color}
\alias{arctic_circle_color}
\title{Arctic Circle line color}
\format{
An object of class \code{character} of length 1.
}
\usage{
arctic_circle_color
}
\description{
Arctic Circle line color
}
\keyword{datasets}
