% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_themes.R
\docType{data}
\name{background_colors}
\alias{background_colors}
\title{Background color palette}
\format{
An object of class \code{character} of length 2.
}
\usage{
background_colors
}
\description{
Default background colors for different projection types
}
\keyword{datasets}
