% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_missing.R
\name{classify_data_issue}
\alias{classify_data_issue}
\title{Classify the type of data issue for a value}
\usage{
classify_data_issue(x)
}
\arguments{
\item{x}{Vector to check}
}
\value{
Character vector with "Missing", "Not reported", "Not relevant", or NA (for valid data)
}
\description{
Classify the type of data issue for a value
}
