% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_join_extraction_data.R
\name{clean_joined_columns}
\alias{clean_joined_columns}
\title{Remove extraneous and duplicate columns}
\usage{
clean_joined_columns(data, columns_to_drop = character())
}
\arguments{
\item{data}{Tibble, the joined data from join_all_literature_modules}

\item{columns_to_drop}{Character vector, names of columns to remove}
}
\value{
A tibble with specified columns removed
}
\description{
Drops specified columns from the joined literature data
}
