% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_node_cards.R
\name{compact_axis_theme}
\alias{compact_axis_theme}
\title{Axis Styling for a Compact Strip}
\usage{
compact_axis_theme()
}
\value{
A ggplot2 theme.
}
\description{
Must be added \strong{after} the complete theme. See \code{\link[=compact_value_scale]{compact_value_scale()}}.
}
