% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_themes.R
\docType{data}
\name{country_colors}
\alias{country_colors}
\title{Country fill color palette}
\format{
An object of class \code{character} of length 2.
}
\usage{
country_colors
}
\description{
Colors for highlighted vs non-highlighted countries
}
\keyword{datasets}
