% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_modify_shapefiles.R
\name{create_graticule}
\alias{create_graticule}
\title{Create graticule lines for map}
\usage{
create_graticule(
  lon_seq = seq(-180, 180, 30),
  lat_seq = seq(60, 90, 10),
  crs = 4326
)
}
\arguments{
\item{lon_seq}{Numeric vector, longitude values for vertical lines}

\item{lat_seq}{Numeric vector, latitude values for horizontal lines}

\item{crs}{CRS object or string, coordinate reference system}
}
\value{
An sf object with graticule lines
}
\description{
Generates latitude and longitude grid lines for map display
}
