% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map_arctic.R
\name{create_study_area_map_wgs84}
\alias{create_study_area_map_wgs84}
\title{Create Arctic study area base map (WGS84 projection)}
\usage{
create_study_area_map_wgs84(
  ocean_sf,
  country_sf,
  arctic_circle_sf,
  graticule_sf = NULL,
  background_color = background_colors["wgs84"],
  bbox = NULL
)
}
\arguments{
\item{ocean_sf}{An sf object with annotated ocean polygons}

\item{country_sf}{An sf object with country polygons}

\item{arctic_circle_sf}{An sf object with Arctic Circle line}

\item{graticule_sf}{An sf object with graticule lines (optional)}

\item{background_color}{Character string, hex color for background}

\item{bbox}{A bounding box used to crop the displayed map. Defaults to none.}
}
\value{
A ggplot object
}
\description{
Generates a ggplot map showing the Arctic study area with oceans,
countries, and the Arctic Circle using WGS84 (EPSG:4326) coordinate system
}
