% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_common_names.R
\name{fetch_common_name_ncbi}
\alias{fetch_common_name_ncbi}
\title{Fetch a Common Name from NCBI via taxize}
\usage{
fetch_common_name_ncbi(sp)
}
\arguments{
\item{sp}{A single scientific name.}
}
\value{
A single string, or \code{NA_character_}.
}
\description{
Fetch a Common Name from NCBI via taxize
}
