% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_triage_species.R
\name{filter_to_species_node}
\alias{filter_to_species_node}
\title{Subset Data to One Species-Group Node}
\usage{
filter_to_species_node(data, node)
}
\arguments{
\item{data}{The \code{literature_analysis_ready} target.}

\item{node}{A one-row tibble from \code{\link[=triage_species_nodes]{triage_species_nodes()}}.}
}
\value{
A filtered data frame.
}
\description{
Subset Data to One Species-Group Node
}
