% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map_arctic.R
\name{geom_sf_shadowtext}
\alias{geom_sf_shadowtext}
\title{Add shadow text to sf plots}
\usage{
geom_sf_shadowtext(
  mapping = aes(),
  data = NULL,
  stat = "sf_coordinates",
  position = "identity",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  check_overlap = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  fun.geometry = NULL
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes()}

\item{data}{The data to be displayed}

\item{stat}{The statistical transformation to use}

\item{position}{Position adjustment}

\item{...}{Other arguments passed to layer}

\item{parse}{If TRUE, labels will be parsed as expressions}

\item{nudge_x}{Horizontal adjustment to nudge labels by}

\item{nudge_y}{Vertical adjustment to nudge labels by}

\item{check_overlap}{If TRUE, overlapping labels will be removed}

\item{na.rm}{If FALSE (default), removes missing values with warning}

\item{show.legend}{Logical indicating whether this layer should be included in legends}

\item{inherit.aes}{If FALSE, overrides default aesthetics}

\item{fun.geometry}{Function for transforming geometry (sf-specific)}
}
\value{
A ggplot2 layer
}
\description{
Custom geom for adding text with shadow/outline to sf plots, useful for
improving label visibility over complex backgrounds
}
