% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_thresholds.R
\name{get_study_area_bbox}
\alias{get_study_area_bbox}
\title{Get study area bounding box}
\usage{
get_study_area_bbox()
}
\value{
An sf bbox object with study area extent
}
\description{
Returns the bounding box coordinates for the Arctic study area
}
