% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_themes.R
\docType{data}
\name{graticule_color}
\alias{graticule_color}
\title{Graticule color}
\format{
An object of class \code{character} of length 1.
}
\usage{
graticule_color
}
\description{
Graticule color
}
\keyword{datasets}
