% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_group_decisions.R
\name{group_decision_human_cols}
\alias{group_decision_human_cols}
\title{Columns Owned by the Human}
\usage{
group_decision_human_cols()
}
\value{
A character vector of column names.
}
\description{
Never overwritten by \code{\link[=scaffold_group_decisions]{scaffold_group_decisions()}}.
}
