% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_missing.R
\name{has_data_issue}
\alias{has_data_issue}
\title{Check if a value has any data quality issue (missing, not reported, or not relevant)}
\usage{
has_data_issue(x)
}
\arguments{
\item{x}{Vector to check}
}
\value{
Logical vector indicating any problematic values
}
\description{
Check if a value has any data quality issue (missing, not reported, or not relevant)
}
