% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_imputation.R
\name{impute_below_limits}
\alias{impute_below_limits}
\title{Impute values below detection/quantification limits}
\usage{
impute_below_limits(
  data,
  measured_col,
  lod_col,
  loq_col,
  impute_fn = function(x) x/sqrt(2),
  output_col = "MEASURED_OR_IMPUTED_VALUE"
)
}
\arguments{
\item{data}{A data frame containing the columns to impute}

\item{measured_col}{Character. Name of column containing measured values}

\item{lod_col}{Character. Name of column containing limit of detection values}

\item{loq_col}{Character. Name of column containing limit of quantification values}

\item{impute_fn}{Function. Function to apply to LOD/LOQ values for imputation.
Default is \code{function(x) x / sqrt(2)}}

\item{output_col}{Character. Name for the output column containing
measured or imputed values. Default is "MEASURED_OR_IMPUTED_VALUE"}
}
\value{
Data frame with added column containing measured or imputed values
}
\description{
Imputes missing values using a specified imputation function, with priority
given to LOD values over LOQ values when both are present.
}
\details{
The function follows this priority order:
\enumerate{
\item Use measured value if available
\item Use imputed LOD value if available
\item Use imputed LOQ value if available
\item Return NA if none available
}
}
\examples{
df <- data.frame(
  measured = c(5, NA, NA),
  lod = c(NA, 2, NA),
  loq = c(NA, NA, 3)
)
impute_below_limits(df, "measured", "lod", "loq")

}
