% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_missing.R
\name{is_missing}
\alias{is_missing}
\title{Check if a value is truly missing (NA, NULL, or empty string)}
\usage{
is_missing(x)
}
\arguments{
\item{x}{Vector to check}
}
\value{
Logical vector indicating missing values
}
\description{
Check if a value is truly missing (NA, NULL, or empty string)
}
