% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_missing.R
\name{is_not_relevant}
\alias{is_not_relevant}
\title{Check if a value is not relevant}
\usage{
is_not_relevant(x)
}
\arguments{
\item{x}{Vector to check}
}
\value{
Logical vector indicating "Not relevant" values
}
\description{
Check if a value is not relevant
}
