% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_missing.R
\name{is_not_reported}
\alias{is_not_reported}
\title{Check if a value is not reported}
\usage{
is_not_reported(x)
}
\arguments{
\item{x}{Vector to check}
}
\value{
Logical vector indicating "Not reported" values
}
\description{
Check if a value is not reported
}
