% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_themes.R
\docType{data}
\name{label_colors}
\alias{label_colors}
\title{Label color palette}
\format{
An object of class \code{character} of length 2.
}
\usage{
label_colors
}
\description{
Colors for different types of labels
}
\keyword{datasets}
