% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_CSVs.R
\name{literature_module_vocab}
\alias{literature_module_vocab}
\title{Literature module vocabulary}
\usage{
literature_module_vocab()
}
\value{
Character vector of module names
}
\description{
Returns standardized names for literature extraction modules
}
