% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_parquets.R
\name{load_literature_parquet}
\alias{load_literature_parquet}
\title{Load literature data from parquet file}
\usage{
load_literature_parquet(
  input_path = "data/clean/derived",
  filename = "literature_data.parquet"
)
}
\arguments{
\item{input_path}{Character string, path where parquet file is located}

\item{filename}{Character string, name of input file (default: "literature_data.parquet")}
}
\value{
A tibble with the literature data
}
\description{
Reads the cleaned, joined literature data from disk in Apache Arrow parquet format
}
