% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_themes.R
\docType{data}
\name{marine_colors}
\alias{marine_colors}
\title{Marine geography color palette}
\format{
An object of class \code{character} of length 2.
}
\usage{
marine_colors
}
\description{
Colors for highlighted vs non-highlighted marine features
}
\keyword{datasets}
