% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_boundaries.R
\name{northern_hemisphere_bbox}
\alias{northern_hemisphere_bbox}
\title{Get Northern Hemisphere bounding box}
\usage{
northern_hemisphere_bbox()
}
\value{
An sf bbox object with Northern Hemisphere extent
}
\description{
Returns the bounding box coordinates for the entire Northern Hemisphere
plus one degree south of the equator
}
