% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_group_notebooks.R
\name{notebook_slug}
\alias{notebook_slug}
\title{Filesystem-Safe Notebook Slug}
\usage{
notebook_slug(notebook)
}
\arguments{
\item{notebook}{A notebook name from \code{\link[=notebook_names]{notebook_names()}}.}
}
\value{
A lower-case, dash-separated slug.
}
\description{
Filesystem-Safe Notebook Slug
}
