% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_occurence_density.R
\name{plot_copper_boxplots}
\alias{plot_copper_boxplots}
\title{Create Boxplot for Copper Measurements from Summary Statistics}
\usage{
plot_copper_boxplots(
  data,
  compartment,
  sub_compartments = NULL,
  thresholds,
  max_threshold = 1
)
}
\arguments{
\item{data}{Filtered measurement data with summary statistics}

\item{compartment}{Environmental compartment to plot}

\item{sub_compartments}{Character vector of sub-compartments to include. Leave NULL to include all}

\item{thresholds}{Threshold data}

\item{max_threshold}{Maximum threshold value to display}
}
\description{
Create Boxplot for Copper Measurements from Summary Statistics
}
