% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_occurence_density.R
\name{plot_copper_ridges_species}
\alias{plot_copper_ridges_species}
\title{Create Ridge Plot for Copper Measurements by Species}
\usage{
plot_copper_ridges_species(
  data,
  compartment = "Biota",
  sub_compartments = NULL,
  thresholds,
  max_threshold = 1,
  n_points = 100,
  scale_ridges = 0.05
)
}
\arguments{
\item{data}{Filtered measurement data}

\item{compartment}{Environmental compartment to plot}

\item{sub_compartments}{Character vector of sub-compartments to include}

\item{thresholds}{Threshold data (from generate_copper_thresholds())}

\item{max_threshold}{Maximum threshold value to display}

\item{n_points}{Number of points for density calculation}

\item{scale_ridges}{Scale parameter for ridgeline heights (default 0.05)}
}
\description{
Create Ridge Plot for Copper Measurements by Species
}
