% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prtr_emissions.R
\name{prtr_facility_files}
\alias{prtr_facility_files}
\title{The PRTR Facility Files and What They Cover}
\usage{
prtr_facility_files()
}
\value{
A named character vector, source category to filename.
}
\description{
The PRTR Facility Files and What They Cover
}
