% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_modify_shapefiles.R
\name{st_line_split}
\alias{st_line_split}
\title{Split sf geometry along a line}
\usage{
st_line_split(geometry, line = c(0, 180), direction = "horizontal")
}
\arguments{
\item{geometry}{An sf geometry object to split}

\item{line}{Numeric vector of length 2, coordinates of the split line}

\item{direction}{Character string, either "horizontal" or "vertical"}
}
\value{
An sf object with the split geometry
}
\description{
Splits an sf geometry object along a horizontal or vertical line by
creating bounding boxes and cropping
}
