% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_themes.R
\name{theme_arctic_map}
\alias{theme_arctic_map}
\title{Arctic map theme}
\usage{
theme_arctic_map(background_color = background_colors["wgs84"], ...)
}
\arguments{
\item{background_color}{Character string, hex color for panel background}

\item{...}{Additional arguments passed to theme()}
}
\value{
A ggplot2 theme object
}
\description{
Custom theme for Arctic maps with minimal styling
}
