% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_modify_shapefiles.R
\name{transform_countries_to_polar}
\alias{transform_countries_to_polar}
\title{Transform country polygons to polar projection}
\usage{
transform_countries_to_polar(countries_wgs84, crs = "EPSG:3575")
}
\arguments{
\item{countries_wgs84}{An sf object with country polygons in WGS84}

\item{crs}{Character string or CRS object, target projection (default: EPSG:3575)}
}
\value{
An sf object with country polygons in polar projection
}
\description{
Transforms country polygons to North Polar Stereographic projection
}
