% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_group_triage.R
\name{triage_group_cols}
\alias{triage_group_cols}
\title{Columns That Define a Sample Group}
\usage{
triage_group_cols()
}
\value{
A character vector of column names.
}
\description{
Columns That Define a Sample Group
}
