% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_units.R
\name{unit_numerator_factors}
\alias{unit_numerator_factors}
\title{Numerator Prefixes, as a Factor to Milligrams}
\usage{
unit_numerator_factors()
}
\value{
A named numeric vector.
}
\description{
Numerator Prefixes, as a Factor to Milligrams
}
