% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_group_triage.R
\name{write_triage_plots}
\alias{write_triage_plots}
\title{Write Triage Plots for Several Groups}
\usage{
write_triage_plots(data, groups, dir = "triage", ...)
}
\arguments{
\item{data}{The \code{literature_analysis_ready} target.}

\item{groups}{Output of \code{\link[=sample_triage_groups]{sample_triage_groups()}}.}

\item{dir}{Output directory.}

\item{...}{Passed to \code{\link[=write_triage_plots_for_group]{write_triage_plots_for_group()}}, notably \code{scale_limits}
and \code{thresholds}.}
}
\value{
A character vector of all written file paths, for \code{format = "file"}.
}
\description{
Write Triage Plots for Several Groups
}
