% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_CSVs.R
\name{fread_all_module_files}
\alias{fread_all_module_files}
\title{Read all files for a given module}
\usage{
fread_all_module_files(file_paths, format_initialiser)
}
\arguments{
\item{file_paths}{A tibble/df of file paths supplied by get_literature_csv_tibble}

\item{format_initialiser}{Function that returns expected tibble structure}
}
\value{
A tibble with all module data combined
}
\description{
Generic function to read and combine all CSV files for a specified module
using the appropriate format initialiser
}
