% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_common_names.R
\name{get_common_names}
\alias{get_common_names}
\title{Get Common Names for Scientific Species Names}
\usage{
get_common_names(
  biota_data,
  input_col = "SAMPLE_SPECIES",
  output_col = "SPECIES_COMMON_NAME",
  cache_path = here::here("data/clean/species_common_names_cache.csv"),
  db = "ncbi",
  verbose = TRUE
)
}
\arguments{
\item{biota_data}{A data frame containing scientific species names}

\item{input_col}{Name of the column containing scientific names. Default is "SAMPLE_SPECIES"}

\item{output_col}{Name of the column to store common names. Default is "SPECIES_COMMON_NAME"}

\item{cache_path}{Path to the cache CSV file. Defaults to "data/clean/species_common_names_cache.csv"}

\item{db}{Database to query for common names. One of "ncbi", "itis", "tropicos", "eol", or "worms". Default is "ncbi"}

\item{verbose}{Logical. Print progress messages? Default is TRUE}
}
\value{
A data frame with common names joined to the input data
}
\description{
Get Common Names for Scientific Species Names
}
\examples{
\dontrun{
biota_with_names <- get_common_names(biota_data)
biota_with_names <- get_common_names(biota_data, db = "worms", verbose = FALSE)
biota_with_names <- get_common_names(
  biota_data,
  input_col = "scientific_name",
  output_col = "common_name"
)
}
}
