% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_join_extraction_data.R
\name{join_all_literature_modules}
\alias{join_all_literature_modules}
\title{Join all literature module tables}
\usage{
join_all_literature_modules(
  measurements_data,
  sites_data,
  reference_data,
  biota_data,
  parameters_data,
  methods_data
)
}
\arguments{
\item{measurements_data}{Tibble, the main fact table}

\item{sites_data}{Tibble, sites dimension table}

\item{reference_data}{Tibble, reference dimension table}

\item{parameters_data}{Tibble, parameters dimension table}

\item{methods_data}{Tibble, methods dimension table (will be spread internally)}

\item{campaign_data}{Tibble, campaign dimension table}
}
\value{
A tibble with all tables joined to measurements
}
\description{
Joins Sites, Reference, Campaign, Parameters, and Methods to the Measurements
fact table using appropriate foreign keys
}
