% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_themes.R
\docType{data}
\name{ocean_colors}
\alias{ocean_colors}
\title{Ocean color palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
ocean_colors
}
\description{
Unnamed vector of colors for specific ocean bodies.
Order matches: Norwegian Sea, Greenland Sea, Barents Sea, Arctic Ocean, default
}
\keyword{datasets}
