% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_units.R
\name{standardise_IDate}
\alias{standardise_IDate}
\title{Standardise a date column to IDate format}
\usage{
standardise_IDate(column, verbose = FALSE, char_format = "dmy")
}
\arguments{
\item{column}{A vector to be converted to IDate. Can be IDate, Date,
character (in dmy or ymd format), or POSIXct.}

\item{verbose}{Logical. If TRUE, prints messages about conversions performed.
Default is FALSE.}

\item{char_format}{Character. The expected format for character dates.
One of "dmy" (default) or "ymd". Only used when column is character.}
}
\value{
An IDate vector
}
\description{
Converts Date, character, or POSIXct columns to data.table's IDate format.
Handles character dates in both dmy and ymd formats.
}
\examples{
standardise_IDate(as.Date("2024-01-15"))
standardise_IDate("15/01/2024", verbose = TRUE)
standardise_IDate("2024-01-15", char_format = "ymd", verbose = TRUE)

}
